Rename *2.tsx to *.tsx
