+++ packages/table/src/headers/abstractHeader.tsx
@@ -18,7 +18,7 @@ import { ILockableLayout, Orientation } from "../interactions/resizeHandle";
 import { DragSelectable, ISelectableProps } from "../interactions/selectable";
 import { ILocator } from "../locator";
 import { IRegion, RegionCardinality, Regions } from "../regions";
-import { IHeaderCellProps } from "./abstractHeaderCell2";
+import { IHeaderCellProps } from "./abstractHeaderCell";
 
 export type IHeaderCellRenderer = (index: number) => React.ReactElement<IHeaderCellProps>;
