+++ packages/table/src/table.tsx
@@ -19,9 +19,9 @@ import { Clipboard } from "./common/clipboard";
 import { Grid } from "./common/grid";
 import { Rect } from "./common/rect";
 import { Utils } from "./common/utils";
-import { ColumnHeader, IColumnWidths } from "./headers/columnHeader2";
-import { ColumnHeaderCell, IColumnHeaderCellProps } from "./headers/columnHeaderCell2";
-import { IRowHeaderRenderer, IRowHeights, renderDefaultRowHeader, RowHeader } from "./headers/rowHeader2";
+import { ColumnHeader, IColumnWidths } from "./headers/columnHeader";
+import { ColumnHeaderCell, IColumnHeaderCellProps } from "./headers/columnHeaderCell";
+import { IRowHeaderRenderer, IRowHeights, renderDefaultRowHeader, RowHeader } from "./headers/rowHeader";
 import { IContextMenuRenderer } from "./interactions/menus";
 import { IIndexedResizeCallback } from "./interactions/resizable";
 import { ResizeSensor } from "./interactions/resizeSensor";
